move FeedForward class into a separated file
